requirements.txt
autots==0.6.6
bcrypt==4.1.2
beautifultable==1.1.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cmdstanpy==1.2.2
contourpy==1.1.1
cryptography==42.0.2
cycler==0.12.1
dnspython==2.5.0
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
fonttools==4.50.0
gunicorn==21.2.0
holidays==0.45
importlib-metadata==7.0.1
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.5
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
patsy==0.5.5
pillow==10.2.0
prophet==1.1.5
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stanio==0.5.0
statsmodels==0.14.1
threadpoolctl==3.2.0
tqdm==4.66.2
tzdata==2023.3
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
